relocation mediaQuery related functions
